add move reduction relation
This is still a work in progress. In particular, while the relation
currently enforces that 1) X goes first 2) no player goes twice in a
row, the relation *does not* enforce the fact that the board needs to
change after a move (i.e. nobody moving is currently a valid reduction
step).